switch to O0

+++ examples/gen-l-video/glv/models/embeddings.py
@@ -25,7 +25,7 @@ def get_timestep_embedding(
     assert len(timesteps.shape) == 1, "Timesteps should be a 1d-array"
 
     half_dim = embedding_dim // 2
-    exponent = -math.log(max_period) * ops.arange(start=0, end=half_dim)
+    exponent = -math.log(max_period) * ops.arange(start=0, end=half_dim, dtype=ms.float32)
     exponent = exponent / (half_dim - downscale_freq_shift)
 
     emb = ops.exp(exponent)
